+++ src/components/Question.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import axios from "axios";
-import { agree, scheduleData, required } from "../questionData";
+import { agree, scheduleData, required } from "../data/questions/questionData";
 import { year } from "../seasonalData";
 
 const Question = () => {
